Last Commit

+++ onlinecounselingbackend/src/main/java/org/cdac/DTO/UserDTO.java
@@ -5,7 +5,7 @@ public class UserDTO {
 	String username;
 	String password;
 	String email;
-	int phone;
+	long phone;
 	String userType;
 	String interestedIn;
 	String fullname;
@@ -33,10 +33,11 @@ public class UserDTO {
 	public void setEmail(String email) {
 		this.email = email;
 	}
-	public int getPhone() {
+	
+	public long getPhone() {
 		return phone;
 	}
-	public void setPhone(int phone) {
+	public void setPhone(long phone) {
 		this.phone = phone;
 	}
 	public String getUserType() {
